Switch boot options location based on distro.
In RaspiOS the boot options are specified in
`/boot/firmware/cmdline.txt`, but in the TPi-provided Ubuntu image for
the RK1 they're in `/boot/firmware/ubuntuEnv.txt`.

There are various ways this could be done, but this is probably the
simplest. We could also split this out into a distro-specific file and
also set `ansible_user` there; at the moment I'm handling this in
`hosts.ini` and `group_vars`.